stash it
